requirements.txt
torch>=0.4.0
torchvision>=0.2.1
dominate>=2.3.1
visdom>=0.1.8.3

pillow~=11.0.0
numpy~=2.1.2
requests~=2.32.3
progressbar~=2.5
matplotlib~=3.9.2
opencv-python~=4.10.0.84
scikit-learn~=1.5.2
scikit-image~=0.24.0